+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **Adverse Event Changes**
   - Adverse event admins can delete closed adverse events
+- **Export Changes**
+  - Exports can no longer be created without selecting an export option
 - **Gem Changes**
   - Update to ruby 2.6.4
   - Update to rails 6.0.1